fix(userdata): Strip \r when checking header
Fix #68
